+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of lizard-nxt client
 Unreleased ()
 -------------
 
--
+- Fix pumpstation and channel new entity types.
 
 Release 0.2b16 (2014-11-26)
 ---------------------------